+++ sdk/src/internal/abstract-private-transfers.ts
@@ -116,6 +116,9 @@ export abstract class AbstractPrivateTransfers implements PrivateTransfersInterf
       transfers: this,
       dappName,
       subAccountAnonymizerAddress: this.subAccountAnonymizerAddress,
+      user: this.user,
+      getViewingKey: () => this.getViewingKey(),
+      discoveryProvider: this.discoveryProvider,
     });
   }